Refactor: 타입 추상화
