+++ src/app/career/types/carouselItem.ts
@@ -1,6 +1,3 @@
-export interface CarouselItem {
-  id: string;
-  title: string;
-  num: string;
-  imageSrc: string;
-}
+// Career 페이지 캐러셀 아이템 타입은 @/app/home/types/project.ts로 이동되었습니다.
+// 하위 호환성을 위해 re-export합니다.
+export type { CarouselItem } from '@/app/home/types';
