+++ src/app/career/types/projectData.ts
@@ -1,19 +1,3 @@
-export interface ProjectData {
-  id: string;
-  title: string;
-  subtitle?: string;
-  num: string;
-  imageSrc: string;
-  period?: string;
-  type?: string;
-  description?: string;
-  role?: string;
-  team?: string;
-  technologies?: string[];
-  achievements?: string[];
-  details?: {
-    problem?: string;
-    solution?: string;
-    impact?: string;
-  };
-}
+// 프로젝트 데이터 타입은 @/app/home/types/project.ts로 이동되었습니다.
+// 하위 호환성을 위해 re-export합니다.
+export type { ProjectData, ProjectDetails } from '@/app/home/types';
