+++ src/app/home/components/traits/CarouselControls.tsx
