+++ src/app/home/components/traits/CarouselNavigation.tsx
@@ -19,32 +19,32 @@ export const CarouselNavigation: React.FC<CarouselNavigationProps> = ({
     <Fragment>
       {/* 하단 도트 네비게이션 */}
       <div className="flex justify-center gap-3 mt-8 pb-4">
-        {Array.from({ length: totalSlides }).map((_, index) => (
+        {Array.from({ length: totalSlides }, (_, index) => (
           <button
             key={index}
             onClick={() => onSlideChange(index)}
-            className={`w-3 h-3 rounded-full transition-all duration-200 ${
+            className={`w-3 h-3 rounded-full transition-all duration-300 ${
               index === currentSlide
-                ? 'bg-white scale-125 shadow-lg'
-                : 'bg-white/40 hover:bg-white/60'
+                ? 'bg-blue-400 scale-125 shadow-lg shadow-blue-400/50'
+                : 'bg-white/30 hover:bg-white/50'
             }`}
+            aria-label={`${index + 1}번째 슬라이드로 이동`}
           />
         ))}
       </div>
 
-      {/* 좌우 네비게이션 버튼 */}
+      {/* 이전/다음 버튼 */}
       <button
         onClick={onPrevSlide}
-        className="absolute left-4 top-1/2 transform -translate-y-1/2 p-4 rounded-full hover:bg-white/25 transition-all duration-200 text-white text-xl backdrop-blur-sm z-20"
-        style={{ display: currentSlide === 0 ? 'none' : 'block' }}
+        className="absolute left-4 top-1/2 -translate-y-1/2 w-10 h-10 rounded-full bg-white/15 hover:bg-white/25 transition-all duration-200 flex items-center justify-center text-white backdrop-blur-sm"
+        aria-label="이전 슬라이드"
       >
         ←
       </button>
-
       <button
         onClick={onNextSlide}
-        className="absolute right-4 top-1/2 transform -translate-y-1/2 p-4 rounded-full hover:bg-white/25 transition-all duration-200 text-white text-xl backdrop-blur-sm z-20"
-        style={{ display: currentSlide === totalSlides - 1 ? 'none' : 'block' }}
+        className="absolute right-4 top-1/2 -translate-y-1/2 w-10 h-10 rounded-full bg-white/15 hover:bg-white/25 transition-all duration-200 flex items-center justify-center text-white backdrop-blur-sm"
+        aria-label="다음 슬라이드"
       >
         →
       </button>
