+++ src/app/home/components/traits/CarouselProgressBar.tsx
@@ -13,14 +13,16 @@ export const CarouselProgressBar: React.FC<CarouselProgressBarProps> = ({
 }) => {
   return (
     <div className="flex-1 mx-4 pt-8">
-      <div className="w-full bg-white/20 rounded-full h-2">
-        <div
-          className="bg-white h-2 rounded-full transition-all duration-300 ease-out shadow-lg"
-          style={{ width: `${((currentSlide + 1) / totalSlides) * 100}%` }}
-        />
-      </div>
-      <div className="flex justify-between mt-3 text-sm text-white/80">
+      <div className="flex items-center gap-2 text-white/70 text-sm">
         <span>{currentSlide + 1}</span>
+        <div className="flex-1 h-1 bg-white/20 rounded-full overflow-hidden">
+          <div
+            className="h-full bg-blue-400 transition-all duration-300 ease-out rounded-full"
+            style={{
+              width: `${((currentSlide + 1) / totalSlides) * 100}%`,
+            }}
+          />
+        </div>
         <span>{totalSlides}</span>
       </div>
     </div>
