+++ src/app/home/components/traits/useCarouselData.ts
@@ -1,35 +1,7 @@
 'use client';
 
 import { useTranslation } from 'react-i18next';
-
-export interface CarouselSlideData {
-  id: string;
-  title: string;
-  subtitle: string;
-  content: string;
-  highlight?: string;
-  closing?: string;
-  features?: Array<{ title: string; content: string }>;
-  activities?: Array<{
-    icon: string;
-    title: string;
-    period: string;
-    description: string;
-  }>;
-  experiences?: Array<{
-    company: string;
-    position: string;
-    period: string;
-    description: string;
-    achievements: string[];
-  }>;
-  contact?: {
-    email: string;
-    github: string;
-    blog: string;
-    resume: string;
-  };
-}
+import { CarouselSlideData } from '@/app/home/types';
 
 export const useCarouselData = () => {
   const { t } = useTranslation();
