+++ src/app/home/types/index.ts
@@ -1,2 +1,8 @@
 // Export all types from trait module
 export * from './trait';
+
+// 캐러셀 관련 타입들
+export * from './carousel';
+
+// 프로젝트 관련 타입들  
+export * from './project';
